app/requirements.txt
aiogram==3.15.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
openai==1.57.4
aiofiles==24.1.0
